+++ minicub3d/set_ray.c
@@ -6,7 +6,7 @@
 /*   By: seojyang <seojyang@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/28 16:07:27 by seojyang          #+#    #+#             */
-/*   Updated: 2023/06/28 18:07:48 by seojyang         ###   ########.fr       */
+/*   Updated: 2023/06/29 12:55:04 by seojyang         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,8 +22,8 @@ t_ray	set_ray(t_info *info, int x)
 	ray.ray_dir[Y] = info->player.dir[Y] + info->player.plane[Y] * cam_x;
 	ray.map[X] = (int)info->player.pos[X];
 	ray.map[Y] = (int)info->player.pos[Y];
-	ray.delta_dist[X] = (int)fabs(1 / ray.ray_dir[X]);
-	ray.delta_dist[Y] = (int)fabs(1 / ray.ray_dir[Y]);
+	ray.delta_dist[X] = fabs(1 / ray.ray_dir[X]);
+	ray.delta_dist[Y] = fabs(1 / ray.ray_dir[Y]);
 	if (ray.ray_dir[X] < 0)
 	{
 		ray.step[X] = -1;
